Refactor Reporting component to remove PDF generation logic and implement HTML-based PDF download; add PDFDownloadButton and PDFReportContent components for better separation of concerns
